She-Balance-App/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
python-multipart==0.0.6
pyjwt==2.8.0
python-decouple==3.8